requirements.txt
flask
sqlalchemy
psycopg2-binary  # Для PostgreSQL
pymysql  # Для MySQL
requests
woocommerce  # Якщо wooapi - це бібліотека для WooCommerce API
redis
flask-sqlalchemy
flask-session
pytz
novaposhta-python-client #бібла новой пошти
typing